refactor(core): extract Blockchain block accessors to own module
Seven impl Blockchain read-side methods move out of blockchain.rs to a
new sibling module crates/sentrix-core/src/blockchain_block_accessors.rs:

  height                — sliding-window position (last block index)
  chain_window_start    — first in-memory block index
  latest_block          — Err on empty chain
  get_block             — in-window lookup, None on miss
  get_block_by_hash     — in-window scan by hash
  get_block_any         — transparent MDBX fallback for evicted blocks
                          (BACKLOG #14, GetBlocks peer handler)
  get_block_reward      — fork-aware halving + remaining-supply cap

Rust permits a split impl Blockchain {} across modules within the same
crate; all bc.method(...) call sites keep resolving unchanged.

Side cleanup: Blockchain::halvings_at (thin wrapper around
crate::tokenomics::halvings_at) is removed — it was crate-private and
the only remaining callers after get_block_reward moved out were five
tests in blockchain::tests, which now call crate::tokenomics::halvings_at
directly. Same value path, slightly more honest about where the math
lives.

blockchain.rs drops from 3406 -> 3289 LOC (-117). 231 / 231 sentrix-core
tests still green; no behaviour change.
